+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Made some chat messages more clear
 * Added `pvphelp <commandName>` to retrieve detailed information about the specified command
 * Restructured the whole `pvphelp` overview menu
+* Global chat messages are now prefixed with a configurable prefix
 
 ### LOTR compatibility:
 * Players now optionally drop their skulls when killed with a weapon with the headhunter modifier even if keepInventory is on